# webserver/angular/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST ",
        "version": "1.0.0",
        "title": "Swagger MyNerdOffice",
        "contact": {
            "email": "saverio.tosi@studenti.unicam.it"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "192.168.1.104:8000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic",
            "description": "HTTP Basic Authentication. Works over `HTTP` and `HTTPS`"
        }
    },
    "paths": {
...
        "/beacon": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Get array of beacons",
                "description": "",
                "operationId": "getBeacons",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Beacon"
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Add a new beacon",
                "description": "",
                "operationId": "postBeacon",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "formData",
                        "description": "The uuid of the beacon.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "major",
                        "in": "formData",
                        "description": "The major id of the beacon.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "minor",
                        "in": "formData",
                        "description": "The minor of the beacon.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
        "/beacon/{id}": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Get beacon by ID",
                "description": "",
                "operationId": "getBeacon",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id that needs to be fetched.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Beacon"
                        }
                    },
                    "401": {
                        "description": "Authorization require"
                    },
                    "404": {
                        "description": "Beacon not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Delete a beacon",
                "description": "",
                "operationId": "deleteBeacon",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the beacon to delete.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
        "/beacon/unregistered": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "tags": [
                    "beacon"
                ],
                "summary": "Get array of unregistered beacons",
                "description": "",
                "operationId": "getUnregistaredBeacons",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Beacon"
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization required"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Beacon": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "minor": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
...
    }
}